requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
httpx==0.27.0
pdfminer.six==20221105
python-multipart==0.0.9
pydantic==2.7.1
python-dotenv==1.0.1
